# docs/swagger.yaml
basePath: /
definitions:
  app.DirectorsList:
    properties:
      date_of_birth:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - date_of_birth
    - name
    type: object
  app.FilmsList:
    properties:
      director_id:
        type: integer
      genre:
        type: string
      id:
        type: integer
      minutes:
        type: number
      name:
        type: string
      rate:
        type: number
      year:
        type: integer
    required:
    - director_id
    - genre
    - minutes
    - name
    - rate
    - year
    type: object
...
  app.UpdateDirectorInput:
    properties:
      date_of_birth:
        type: string
      name:
        type: string
    type: object
  app.UpdateFilmInput:
    properties:
      direcor_id:
        type: integer
      genre:
        type: string
      minutes:
        type: number
      name:
        type: string
      rate:
        type: number
      year:
        type: integer
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.getAllDirectorsResponce:
    properties:
      directors:
        items:
          $ref: '#/definitions/app.DirectorsList'
        type: array
    type: object
...
  handler.getAllFilmsResponce:
    properties:
      films:
        items:
          $ref: '#/definitions/app.FilmsList'
        type: array
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API Server for FilmsList Application
  version: "1.0"
paths:
  /api-admin/directors/{id}:
    delete:
      consumes:
      - application/json
      description: delete director
      operationId: delete-director
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete director
      tags:
      - admin api directors
    get:
      consumes:
      - application/json
      description: get director by id
      operationId: get-director-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.DirectorsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Director By ID
      tags:
      - admin api directors
    put:
      consumes:
      - application/json
      description: update director
      operationId: update-director
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: director info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.UpdateDirectorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update director
      tags:
      - admin api directors
  /api-admin/directors/all:
    get:
      consumes:
      - application/json
      description: get all directors
      operationId: get-all-directors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllDirectorsResponce'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Directors
      tags:
      - admin api directors
  /api-admin/directors/create:
    post:
      consumes:
      - application/json
      description: create director
      operationId: create-director
      parameters:
      - description: director info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.DirectorsList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create director
      tags:
      - admin api directors
  /api-admin/films/{id}:
    delete:
      consumes:
      - application/json
      description: delete film
      operationId: delete-film
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: film info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.UpdateFilmInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete film
      tags:
      - admin api films
    get:
      consumes:
      - application/json
      description: get film by id
      operationId: get-film-by-id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.FilmsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Film By ID
      tags:
      - admin api films
    put:
      consumes:
      - application/json
      description: update film
      operationId: update-film
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: film info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.UpdateFilmInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update film
      tags:
      - admin api films
  /api-admin/films/all:
    get:
      consumes:
      - application/json
      description: get all films
      operationId: get-all-films
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllFilmsResponce'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Films
      tags:
      - admin api films
  /api-admin/films/create:
    post:
      consumes:
      - application/json
      description: create film
      operationId: create-film
      parameters:
      - description: film info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.FilmsList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create film
      tags:
      - admin api films
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"